[Patchset 5] perfetto-ui: Add cpu freq tracks
The MVP for cpu freq tracks - adds a track for each cpu
if we have cpu freq events for that cpu in the trace.
Also center text in tool tip.

Bug:118898348

+++ ui/src/tracks/cpu_freq/common.ts
@@ -22,7 +22,7 @@ export interface Data {
   minimumValue: number;
 
   timestamps: Float64Array;
-  valuesKHz: Float32Array;
+  valuesKHz: Uint32Array;
 }
 
 export interface Config {
